+++ stroom-query/stroom-query-common/src/main/java/stroom/query/common/v2/EventCoprocessor.java
@@ -26,7 +26,7 @@ import com.esotericsoftware.kryo.io.Output;
 import java.util.List;
 import java.util.concurrent.locks.ReentrantLock;
 
-public class EventCoprocessor implements Coprocessor {
+public class EventCoprocessor implements Coprocessor, HasCompletionState {
 
     private static final String STREAM_ID = "StreamId";
     private static final String EVENT_ID = "EventId";